renamed processor to ecslogattributes
