+++ otel-agent/ecs-ec2/ecslogattributesprocessor/config.go
@@ -1,4 +1,4 @@
-package ecslogresourcedetectionprocessor
+package ecslogattributesprocessor
 
 import (
 	"regexp"
