+++ otel-agent/ecs-ec2/ecslogattributesprocessor/config_test.go
@@ -1,4 +1,4 @@
-package ecslogresourcedetectionprocessor
+package ecslogattributesprocessor
 
 import (
 	"path/filepath"
